Refactor: utils/geometry에 원의 좌표 계산 함수 분리
